chore(labels): rename label sync entrypoint
