+++ scripts/test-labels-dry-run.sh
@@ -2,7 +2,8 @@
 set -eu
 
 ROOT=$(CDPATH= cd -- "$(dirname -- "$0")/.." && pwd)
-SCRIPT="$ROOT/scripts/labels-dry-run.rb"
+SCRIPT="$ROOT/scripts/labels-sync.rb"
+COMPAT_SCRIPT="$ROOT/scripts/labels-dry-run.rb"
 TMPDIR=${TMPDIR:-/tmp}
 TEST_TMP=$(mktemp -d "$TMPDIR/labels-dry-run-test.XXXXXX")
 OUT="$TEST_TMP/out"
@@ -42,6 +43,7 @@ assert_json_field() {
 }
 
 ruby -c "$SCRIPT" >/dev/null
+ruby -c "$COMPAT_SCRIPT" >/dev/null
 
 # Existing refusal paths.
 assert_exit 2 "$SCRIPT"
@@ -54,7 +56,7 @@ assert_exit 2 "$SCRIPT" --all-repos --apply
 assert_exit 2 "$SCRIPT" --all-repos --apply --confirm-apply
 assert_exit 2 "$SCRIPT" --repo z-shell/zi --apply --confirm-apply
 
-# Apply preview must be machine-readable and non-mutating.
-assert_json_field apply-preview "$SCRIPT" --repo z-shell/.github --apply --json
+# Backward-compatible dry-run entrypoint should keep working.
+assert_json_field apply-preview "$COMPAT_SCRIPT" --repo z-shell/.github --apply --json
 
-printf 'labels-dry-run smoke tests passed\n'
+printf 'labels-sync smoke tests passed\n'
